+++ src/odes/iode.jl
@@ -337,7 +337,7 @@ end
 
 function compute_vectorfields!(state::State, prob::IODEProblem)
     initialguess(prob).v(state.q̇, state.t, state.q, state.p, parameters(prob))
-    initialguess(prob).f(state.ṗ, state.t, state.q, state.q̇, parameters(prob))
+    initialguess(prob).f(state.ṗ, state.t, state.q, state.q̇, parameters(prob))
 end
 
 @doc """