+++ src/sdes/psde.jl
@@ -270,7 +270,7 @@ end
 
 function compute_vectorfields!(state::State, prob::PSDEProblem)
     initialguess(prob).v(state.q̇, state.t, state.q, state.p, parameters(prob))
-    initialguess(prob).f(state.ṗ, state.t, state.q, state.p, parameters(prob))
+    initialguess(prob).f(state.ṗ, state.t, state.q, state.p, parameters(prob))
 end
 
 const PSDEEnsemble = EnsembleProblem{PSDE}